ul-30 Updated test_registration
Moved test for set password api endpoint into test_registration.py
module. This module contains the entire test suite for registration logic
